+++ apps/wpcom-block-editor/src/calypso/features/iframe-bridge-server.js
@@ -19,7 +19,7 @@ import { Component, useEffect, useState } from 'react';
 import tinymce from 'tinymce/tinymce';
 import { STORE_KEY as NAV_SIDEBAR_STORE_KEY } from '../../../../editing-toolkit/editing-toolkit-plugin/wpcom-block-editor-nav-sidebar/src/constants';
 import { inIframe, isEditorReadyWithBlocks, sendMessage, getPages } from '../../utils';
-import FeedbackForm from './feedback-form';
+import FeedbackForm from './fse-beta/feedback-form';
 /**
  * Conditional dependency.  We cannot use the standard 'import' since this package is
  * not available in the post editor and causes WSOD in that case.  Instead, we can
